+++ drift/CHANGELOG.md
@@ -11,6 +11,7 @@
   inside of the transaction, instead of causing a deadlock.
 - Improve stack traces for errors happening on drift isolates (which includes
   usages of `NativeDatabase.createInBackground`).
+- Don't cache `EXPLAIN` statements, avoiding schema locks.
 
 ## 2.15.0